(PC-33579) refactor(CategoriesButtonDisplay): move header into its own file
